requirements.txt
flask
opencv-python-headless
pyzbar
pillow
flask-cors
gunicorn